Updated sendEmail.php. Added isset test and sticky fields
added if statement with isset clause to check if the submit value had been set to distinguish the event between displaying the form the first time or submitting an email with blank fields. Added echo statements to the value attribute of the  subject field and to the text area to make the form sticky.